requirements_txt.txt
discord.py==2.3.2
aiohttp==3.8.6
python-dotenv==1.0.0
asyncio
sqlite3
requests==2.31.0
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
notion-client==2.0.0